Modify some files and create Main class to run app

+++ src/com/sda/billApp/CoinEur1.java
@@ -6,7 +6,7 @@ public class CoinEur1 extends Coin{
     }
 
     @Override
-    public int getValue() {
+    public double getValue() {
         return 1;
     }
